geo-tracker/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Redis & Caching
redis==5.0.1
celery==5.3.4

# AI & ML
openai==1.3.7
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# HTTP & Scraping
httpx==0.25.2
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Authentication
python-jose[cryptography]==3.3.0
cryptography==41.0.7
PyJWT[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# JWKS
joserfc==0.9.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.0
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI
faker==20.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1